Update language configuration
This adds atoms, strings and binaries the pairs definitions, and adds
indentation rules. These are not perfect, but the best you can do
within the limitations of VS Code.

Especially `end` is hard, as it commonly should be dedented twice,
which you can't specify using language-configuration.json. Even if
you could, sometimes you shouldn't dedent twice, so it wouldn't be
perfect either.

Still, this is a big improvement in ergonomics.